Add not found page

+++ client/src/components/detailsPageComponent/Details.jsx
@@ -8,19 +8,25 @@ import HeaderDetails from "./HeaderDetails";
 import Comments from './Comments';
 
 export default function Details() {
-    const {estateId} = useParams(); 
-    const [estate,setEstate] = useState({})
+    const { estateId } = useParams();
+    const [estate, setEstate] = useState({})
     const navigate = useNavigate();
-    useEffect(()=>{
+    useEffect(() => {
         estateService.getOne(estateId)
             .then(setEstate)
-            .catch(err=> console.log(err))
-    },[estateId])
-    
+            .catch(err => {
+                console.log(err)
+                navigate('/')
+            })
+    }, [estateId])
+
     const onDeleteClickHandler = async (estateId) => {
         try {
-            await estateService.remove(estateId);
-            navigate('/catalog')
+            const hasConfirmed = confirm(`Are you sure you want to delete ${estate.title}`);
+            if (hasConfirmed) {
+                await estateService.remove(estateId);
+                navigate('/catalog')
+            }
         } catch (error) {
             console.log(error);
         }
@@ -28,10 +34,10 @@ export default function Details() {
     return (
         <>
             <HeaderDetails />
-            
-            <DetailInformation {...estate} onDeleteClickHandler={onDeleteClickHandler}/>
 
-            <Comments {...estate}/>
+            <DetailInformation {...estate} onDeleteClickHandler={onDeleteClickHandler} />
+
+            <Comments {...estate} />
         </>
     )
 }
